Some new functions implemented
used awesome notification for flutter. updated androidmenifest.

the main issue was that i didn't upgrade my dependencies. so it was creating issue with android 13 as the lower versions of my dependencies were built for lower than android 12 & 13 version.

also used geolocator to open location settings if it gets location is turned off after entering mapTracker page. and after that if user still doesn't turn on location then it will log out the user and take user back to homepage. this functionality is implemented at didChangeAppLifecycleState method. this functionality works when app resumes as when app pass the user to location settings it gets paused and then when user comes back it gets resumed so there we implemented the functionalities.

Pushed notification when user starts tracking & after user log out.

notification controller is useless. Will remove it in the next version.

for awesome notification to work fine we had to change the minsdk version from 19 to 21.

nothing important was changed in main.dart

and no ui design was changed in this version.